+++ opencascade-hs/CHANGELOG.md
@@ -8,7 +8,13 @@ and this project adheres to the
 
 ## Unreleased
 
+### Added 
+
+- Added OpenCascade.BRepPrimAPI.MakeCone
+
 ## 0.1.1.1 - 2023-12-12 
+
+### Fixed 
 
 - Include header files in package, fixing build